Continue Fixing test execution
